+++ src/main/java/br/comexport/avaliacao/services/QuestionService.java
@@ -1,11 +1,11 @@
 package br.comexport.avaliacao.services;
 
-import br.comexport.avaliacao.entities.FlagsEntity;
-import br.comexport.avaliacao.entities.QuestionEntity;
-import br.comexport.avaliacao.entities.UserEntity;
+import br.comexport.avaliacao.entities.*;
 import br.comexport.avaliacao.exception.ResourceNotFoundException;
 import br.comexport.avaliacao.parameters.QuestionParameter;
+import br.comexport.avaliacao.repositories.AnswerRepository;
 import br.comexport.avaliacao.repositories.QuestionRepository;
+import br.comexport.avaliacao.repositories.VoteAnswerRepository;
 import br.comexport.avaliacao.util.Util;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
@@ -18,6 +18,12 @@ public class QuestionService {
 
     @Autowired
     QuestionRepository questionRepository;
+
+    @Autowired
+    AnswerRepository answerRepository;
+
+    @Autowired
+    VoteAnswerRepository voteAnswerRepository;
 
     @Autowired
     Util util;
@@ -48,8 +54,23 @@ public class QuestionService {
     }
 
     public ResponseEntity<?> deleteQuestion(Long id){
+
         QuestionEntity question = questionRepository.findById(id)
                 .orElseThrow(() -> new ResourceNotFoundException("Question", "id", id));
+
+        List<AnswerEntity> answerEntities = answerRepository.selectAnswerQuestion(id);
+
+        if(answerEntities.size() > 0) {
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST)
+                    .body(util.exceptError("Dados inválidos","Não é possivel excluir pergunta sem antes excluir a resposta!"));
+        }
+
+        List<VoteAnswerEntity> voteAnswerEntities = voteAnswerRepository.selectQuestion(id);
+
+        if( voteAnswerEntities.size() > 0) {
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST)
+                    .body(util.exceptError("Dados inválidos","Não é possivel excluir pergunta sem antes excluir o voto da resposta!"));
+        }
 
         questionRepository.delete(question);
 
@@ -68,6 +89,7 @@ public class QuestionService {
         questionEntity.setComment(questionParameter.getComment());
         questionEntity.setFlag(new FlagsEntity(questionParameter.getId_flag()));
         questionEntity.setUser(new UserEntity(questionParameter.getId_user()));
+        questionEntity.setId(id);
         questionEntity.setResolved(questionParameter.isResolved());
 
         questionRepository.save(questionEntity);
@@ -81,12 +103,12 @@ public class QuestionService {
                     .body(util.exceptError("Dados inválidos","Comentário não informado!"));
         }
 
-        if ( questionParameter.getId_flag().intValue() < 1 ) {
+        if ( questionParameter.getId_flag().longValue() < 1 ) {
             return ResponseEntity.status(HttpStatus.BAD_REQUEST)
                     .body(util.exceptError("Dados inválidos","Flag não informado!"));
         }
 
-        if ( questionParameter.getId_user().intValue() < 1 ) {
+        if ( questionParameter.getId_user().longValue() < 1 ) {
             return ResponseEntity.status(HttpStatus.BAD_REQUEST)
                     .body(util.exceptError("Dados inválidos","Usuário não informado!"));
         }
